pyproject.toml
[tool.poetry]
name = "aero-bloodchain"
version = "0.1.0"
description = "熱血飛騰：血品供應韌性系統 - Emergency Blood Chain System"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.31.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
langchain-community = "^0.0.20"
openai = "^1.10.0"
sqlalchemy = "^2.0.25"
pandas = "^2.2.0"
folium = "^0.15.1"
plotly = "^5.18.0"
matplotlib = "^3.8.2"
seaborn = "^0.13.1"
sqlite-vec = "^0.1.1"
python-dotenv = "^1.0.1"
reportlab = "^4.0.9"
openpyxl = "^3.1.2"
xlsxwriter = "^3.1.9"
streamlit-folium = "^0.17.0"
geopy = "^2.4.1"
scikit-learn = "^1.4.0"
langchain-experimental = "^0.0.50"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^24.1.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

